style: 차트관련 파일, 폴더들을 CoinDetails 하위로 이동
