--- CryptocurrencyExchange/CryptocurrencyExchange/CoinDetails/ChildViewControllers/Chart/ChartViewController.swift
@@ -1,18 +1,0 @@
-//
-//  ChartViewController.swift
-//  CryptocurrencyExchange
-//
-//  Created by Dayeon Jung on 2022/03/03.
-//
-
-import UIKit
-
-class ChartViewController: ViewControllerInjectingViewModel<ChartsViewModel> {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-
-        // Do any additional setup after loading the view.
-    }
-
-}
