--- CryptocurrencyExchange/CryptocurrencyExchange/CoinDetails/ChildViewControllers/Chart/ChartsViewModel.swift
@@ -1,16 +1,0 @@
-//
-//  ChartsViewModel.swift
-//  CryptocurrencyExchange
-//
-//  Created by Dayeon Jung on 2022/03/03.
-//
-
-import Foundation
-
-class ChartsViewModel: XIBInformation {
-    var nibName: String?
-    
-    init(nibName: String?) {
-        self.nibName = nibName
-    }
-}
