+++ CryptocurrencyExchange/CryptocurrencyExchange/CoinDetails/ChildViewControllers/ChartsByCoin/CandleStickAPI/CandleStickAPIManager.swift
