+++ CryptocurrencyExchange/CryptocurrencyExchange/CoinDetails/ChildViewControllers/ChartsByCoin/CandleStickAPI/Entity/CandleStickEntity.swift
