+++ CryptocurrencyExchange/CryptocurrencyExchange/CoinDetails/ChildViewControllers/ChartsByCoin/ChartByTimes/ChartView/CandleStickChart/Model/CandleChartDataEntries.swift
