+++ CryptocurrencyExchange/CryptocurrencyExchange/CoinDetails/ChildViewControllers/ChartsByCoin/ChartByTimes/ChartView/CandleStickChart/View/CandleStickChartByTimesView.xib
