+++ CryptocurrencyExchange/CryptocurrencyExchange/CoinDetails/ChildViewControllers/ChartsByCoin/ChartByTimes/ChartView/CandleStickChart/ViewModel/CandleStickChartByTimesViewModel.swift
