+++ CryptocurrencyExchange/CryptocurrencyExchange/CoinDetails/ChildViewControllers/ChartsByCoin/ChartByTimes/CoinInformation/Model/SignSymbol.swift
