+++ CryptocurrencyExchange/CryptocurrencyExchange/CoinDetails/ChildViewControllers/ChartsByCoin/ChartByTimes/CoinInformation/View/CoinInformationInChartView.swift
