+++ CryptocurrencyExchange/CryptocurrencyExchange/CoinDetails/ChildViewControllers/ChartsByCoin/ChartByTimes/Model/IntervalButton.swift
