+++ CryptocurrencyExchange/CryptocurrencyExchange/CoinDetails/ChildViewControllers/ChartsByCoin/ChartByTimes/ViewModel/ChartByTimesViewModel.swift
