+++ CryptocurrencyExchange/CryptocurrencyExchange/CoinDetails/ChildViewControllers/ChartsByCoin/View/CoinDetailViewController.xib
